Stop trying to get by on our good looks and charm (or at least stop trying to assume that x8664 Darwin's malloc() will quickly return a 32-bit pointer in allocate_tcr().  See ticket:1005.) Instead, try to map a largish (1K) number of TCRs in free 32-bit memory and manage them explicitly on x8664 Darwin.
Note that this isn't thread-safe in general: we do this by walking
our address space (via vm_region_64()) until we find a free 32-bit
block of memory and using mmap() (with the MAP_FIXED option).  When
this happens at any time after application startup, it's possible
for some foreign thread to be mapping/unmapping regions while we're
doing this.  (This is why OSes that provide mmap options that request
32-bit addresses do so in the kernel.)  It's likely fairly hard in
practice to exceed the 1K initial TCR allocation and it's not clear
that this is any worse than the "wait until we get lucky with malloc()"
strategy has been, but it may be better to just do the TCR allocation
once on startup, avoid the (theoretical) thread-safety issues, and
treat the (possibly raised) value of TCR_CLUSTER_COUNT as a hard limit.

lisp-kernel/platform-darwinx8664: define DARWIN64, to make conditionalization
a little easier
lisp-kernel/memory.c: implement darwin_allocate_tcr() and darwin_free_tcr()
as outlined above
lisp-kernel/thread_manager.c: allocate_tcr() uses darwin_allocate_tcr() on
DARWIN64.  Use darwin_free_tcr() instead of free() on DARWIN64.  Make
shutdown_thread_tcr() dequeue the TCR and put it back in the free TCR
pool on DARWIN64.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15437 8bdb8f1d-59ce-db11-9e41-0016172a54ae