When "eliding pushes", if the pushed and popped registers are disjoint and the popped register is set in the intervening sequence but the pushed register isn't we need to generate a COPY of the pushed reg to the popped reg at the point of the POP.
The x86 backend got this wrong in both the FPR and GPR cases; the ARM
got this right in the GPR case but missed an FPR case.

Fixes ticket:1037 in the trunk; the bug is about a year old, but the
code in question has only been used for PROG1 in the trunk for the last
week or so.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15542 8bdb8f1d-59ce-db11-9e41-0016172a54ae