Define, export, and provide doc strings for DEFINE-CHARACTER-ENCODING-ALIAS and REMOVE-CHARACTER-ENCODING-ALIAS.
When parsing the "coding:" file option, warn (don't error) if the option's
value isn't recognized (todo: handle this warning in the Cocoa IDE) and
suggest use of DEFINE-CHARACTER-ENCODING-ALIAS in the warning.

Define Emacs-compatible aliases for :LATINn (:LATIN-n, :ISO-LATIN-n) and 
:UTF-8 (:MULE-UTF-8).

Fixes ticket:1038 in the trunk.  (Failure to recognize "latin-1" apparently
prevents some part of SLIME or SWANK from loading properly.)


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15545 8bdb8f1d-59ce-db11-9e41-0016172a54ae