arm-exceptions.c, x86-exceptions.c, ppc-exceptions.c: if we set an  egc refbit, set ephemeral_ref. memory.c: clear ephermal_ref when zeroing refbits if managed_static_area  empty.  Screw: need two booleans. gc.h, gc-common.c: mark_memoized_area returns a boolean.  Only  process (mark-,forward-) memoized area if ephemeral_ref is true.
As indicated above: if this idea (only process the refmap if there's
reason to believe that it contains an ephemeral reference) is worthwhile,
it's likely to be a better idea to maintain two flags in the write barrier
on platforms where we use a managed static area.

For now, see if this has the intended effect.


git-svn-id: http://svn.clozure.com/publicsvn/openmcl/trunk/source@15814 8bdb8f1d-59ce-db11-9e41-0016172a54ae